return a better error message if name is taken
Previously, we were returning a 500 error if a baseline's name was
already taken during a PATCH.

Instead, return a 400 with error message. We already had a unit test
for this but the test was (correctly) returning a 400 because I was
sending in a malformed request. The test has been adjusted to test for
the correct error scenario.