fix: unbreak delete by ID list
The body on DELETE requests was not being passed in anymore. I am not
yet sure why that is.

This commit reads the body off of the request instead of using
connexion's parameter passing. A future commit will fix this to use
connexion again.

This also pins werkzeug as 0.16 for now.