fix(floorist): RHICOMPL-3061 explicitly set the ID to be textual
When it is not possible to determine the type of a UUID column from the
first line of a query result because it is empty, the wrangler can fail
on saving the parquet file as it doesn't support UUID fields by default.

As a workaround for this LEFT OUTER JOIN, we can set the field to be
converted to text. This is what the automatic conversion does anyway.